Move device UI to a separate file.
